perf: optimize rule engine hot path with lazy hydration and pre-parsed cache
- Centralize SafeFrappeAPI and evaluation helpers in runtime_eval.py to reduce object churn
- Implement Lazy Rule Hydration in RuleCoordinator to avoid expensive document hydration for ineligible rules
- Expand Action Plan Cache to pre-parse JSON configurations for all core action types
- Refactor RuleEngine and ActionHandlers to leverage pre-parsed plans and unified context
- Move HandlerRegistry initialization out of the execution loop
