+++ flexirule/ruleflow/core/engine.py
@@ -31,7 +31,6 @@ from flexirule.ruleflow.core.contracts import (
 	is_release_disabled_action,
 	normalize_action_type,
 )
-from flexirule.ruleflow.core.evaluator import check_link_match
 from flexirule.ruleflow.core.exceptions import (
 	CycleDetectedError,
 	EmptyRuleError,
@@ -39,13 +38,15 @@ from flexirule.ruleflow.core.exceptions import (
 	RuleDisabledError,
 )
 from flexirule.ruleflow.core.exceptions import TimeoutError as FlexiRuleTimeoutError
-from flexirule.ruleflow.core.runtime_eval import eval_condition_bool, eval_value
-from flexirule.ruleflow.utils.field_resolver import FieldResolver
-from flexirule.ruleflow.utils.mapping import apply_output_mapping
-from flexirule.ruleflow.utils.schema_validator import (
-	frappe_fields_to_json_schema,
-	get_custom_validator,
+from flexirule.ruleflow.core.runtime_eval import (
+	ReadOnlyDocument,
+	SafeFrappeAPI,
+	_safe_frappe,
+	eval_condition_bool,
+	eval_value,
+	get_base_eval_context,
 )
+from flexirule.ruleflow.utils.mapping import apply_output_mapping
 
 # Configuration constants
 MAX_SUB_RULE_DEPTH = 2  # Maximum nesting depth for sub-rule calls
@@ -55,55 +56,6 @@ class TimeoutException(Exception):
 	"""Internal timeout exception"""
 
 	pass
-
-
-class ReadOnlyDocument:
-	"""Proxy for Document that prevents mutation"""
-
-	def __init__(self, doc):
-		object.__setattr__(self, "_doc", doc)
-
-	def __getattr__(self, name):
-		return getattr(self._doc, name)
-
-	def __getitem__(self, key):
-		return self._doc[key]
-
-	def get(self, key, default=None):
-		return self._doc.get(key, default)
-
-	def __setattr__(self, name, value):
-		raise frappe.ValidationError("Cannot mutate document in Pure method")
-
-	def __setitem__(self, key, value):
-		raise frappe.ValidationError("Cannot mutate document in Pure method")
-
-	def save(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot save document in Pure method")
-
-	def insert(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot insert document in Pure method")
-
-	def delete(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot delete document in Pure method")
-
-	def db_set(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot db_set document in Pure method")
-
-	def run_method(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot run_method on document in Pure method")
-
-	def add_comment(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot add_comment on document in Pure method")
-
-	def queue_action(self, *args, **kwargs):
-		raise frappe.ValidationError("Cannot queue_action on document in Pure method")
-
-	@property
-	def flags(self):
-		import copy
-
-		return copy.deepcopy(self._doc.flags)
 
 
 @contextmanager
@@ -128,114 +80,6 @@ def time_limit(seconds):
 		timer.cancel()
 
 
-# TODO : Explore frappe source code to find if they have a similar implementation and use it instead Introduced for SafeFrappeAPI
-class SafeFrappeAPI:
-	"""
-	Restricted Frappe API proxy for rule condition evaluation.
-	Exposes only safe, read-only operations to prevent security issues.
-	"""
-
-	def __init__(self):
-		# Safe utilities
-		self.utils = frappe.utils
-		self._dict = frappe._dict
-
-	@property
-	def session(self):
-		"""Read-only access to frappe.session (current user, roles, etc.)"""
-		return frappe.session
-
-	@staticmethod
-	def get_roles(user=None):
-		"""Read-only: return roles for the given user (or current session user)."""
-		return frappe.get_roles(user)
-
-	# Safe read operations
-	@staticmethod
-	def get_value(doctype, filters, fieldname=None, **kwargs):
-		"""Read-only get_value"""
-		return frappe.get_value(doctype, filters, fieldname, **kwargs)
-
-	@staticmethod
-	def get_all(doctype, filters=None, fields=None, limit_page_length=500, **kwargs):
-		"""Read-only get_all with default limit"""
-		if "limit" in kwargs:
-			limit_page_length = kwargs.pop("limit")
-		return frappe.get_all(
-			doctype, filters=filters, fields=fields, limit_page_length=limit_page_length, **kwargs
-		)
-
-	@staticmethod
-	def db_exists(doctype, name):
-		"""Check if document exists"""
-		return frappe.db.exists(doctype, name)
-
-	@staticmethod
-	def get_meta(doctype):
-		"""Get doctype metadata"""
-		return frappe.get_meta(doctype)
-
-	@staticmethod
-	def format_value(value, df=None, doc=None, currency=None):
-		"""Format value for display"""
-		return frappe.format_value(value, df, doc, currency)
-
-	# Logging (safe)
-	@staticmethod
-	def log(message):
-		"""Log a message"""
-		frappe.logger().info(message)
-
-	# Explicitly denied operations (will raise)
-	def get_doc(self, *args, **kwargs):
-		raise PermissionError("get_doc is not allowed in rule conditions. Use frappe.get_value instead.")
-
-	def new_doc(self, *args, **kwargs):
-		raise PermissionError("new_doc is not allowed in rule conditions.")
-
-	def delete_doc(self, *args, **kwargs):
-		raise PermissionError("delete_doc is not allowed in rule conditions.")
-
-	def db_set_value(self, *args, **kwargs):
-		raise PermissionError("db.set_value is not allowed in rule conditions.")
-
-	@property
-	def db(self):
-		"""Return restricted db proxy"""
-		return self._SafeDB()
-
-	class _SafeDB:
-		"""Restricted database operations"""
-
-		def exists(self, doctype, name):
-			return frappe.db.exists(doctype, name)
-
-		def get_value(self, doctype, filters, fieldname=None, **kwargs):
-			return frappe.db.get_value(doctype, filters, fieldname, **kwargs)
-
-		def get_all(self, doctype, filters=None, fields=None, **kwargs):
-			return frappe.db.get_all(doctype, filters=filters, fields=fields, **kwargs)
-
-		# Explicitly deny write operations
-		def set_value(self, *args, **kwargs):
-			raise PermissionError("db.set_value is not allowed in rule conditions.")
-
-		def sql(self, *args, **kwargs):
-			raise PermissionError("db.sql is not allowed in rule conditions.")
-
-		# Transaction control restricted (following Frappe restrict_commit_rollback)
-		def commit(self, *args, **kwargs):
-			raise PermissionError("db.commit is not allowed during doc event rules.")
-
-		def rollback(self, *args, **kwargs):
-			raise PermissionError("db.rollback is not allowed during doc event rules.")
-
-		def add_index(self, *args, **kwargs):
-			raise PermissionError("db.add_index is not allowed during doc event rules.")
-
-
-# Singleton instance
-_safe_frappe = SafeFrappeAPI()
 
 
 class RuleEngine:
@@ -273,9 +117,14 @@ class RuleEngine:
 		self.action_map_by_label = {a.action_label: a for a in self.actions}
 		self.action_map_by_name = {a.name: a for a in self.actions}
 
-	@staticmethod
-	def _get_action_config(action):
-		"""Get config JSON from action"""
+	def _get_action_config(self, action):
+		"""Get pre-parsed config JSON from action plan cache."""
+		from flexirule.ruleflow.core.action_plan_cache import get_action_plan
+
+		plan = get_action_plan(self.rule, action)
+		if plan and "config" in plan:
+			return plan["config"]
+
 		config_str = getattr(action, "config", None)
 		if not config_str:
 			return {}
@@ -337,6 +186,10 @@ class RuleEngine:
 			timeout = self.rule.max_execution_time or 30
 			context["_timeout"] = timeout
 			context["_start_time"] = start_time
+
+			from flexirule.ruleflow.core.action_handlers import HandlerRegistry
+
+			HandlerRegistry._ensure_initialized()
 
 			if self.context.get("test_mode"):
 				# No timeout in test mode
@@ -521,7 +374,7 @@ class RuleEngine:
 			)
 
 			try:
-				# Use Strategy Pattern with Handler Registry
+				# Use Strategy Pattern with Handler Registry (already initialized)
 				from flexirule.ruleflow.core.action_handlers import HandlerRegistry
 
 				handler = HandlerRegistry.get(normalize_action_type(current.action_type))
@@ -551,6 +404,7 @@ class RuleEngine:
 					except Exception:
 						self.path_trace[-1]["output"] = str(result)
 					try:
+						# Inputs are now pre-parsed in the plan
 						self.path_trace[-1]["input"] = self._get_action_config(current)
 					except Exception:
 						pass
@@ -711,11 +565,11 @@ class RuleEngine:
 	def _build_eval_locals(self, context):
 		"""Build safe locals for expression evaluation."""
 		doc = context.get("doc")
-		if isinstance(doc, dict):
-			doc = frappe._dict(doc)
 		old_doc = context.get("old_doc")
-		if isinstance(old_doc, dict):
-			old_doc = frappe._dict(old_doc)
+		vars_dict = context.get("vars", {})
+
+		eval_locals = get_base_eval_context(doc, old_doc=old_doc, vars_dict=vars_dict)
+
 		rule_meta = context.get("rule") or {
 			"name": self.rule.name,
 			"trigger_type": self.rule.trigger_type,
@@ -731,73 +585,25 @@ class RuleEngine:
 				"trigger_event": meta_ctx.get("caller_trigger_event"),
 				"document_type": meta_ctx.get("caller_document_type"),
 			}
+
 		doctype_name = (
 			context.get("doctype")
 			or rule_meta.get("document_type")
-			or getattr(doc, "doctype", None)
+			or (getattr(doc, "doctype", None) if doc else None)
 			or self.rule.document_type
 		)
 
-		def _get_meta(doctype):
-			if not doctype:
-				return None
-			try:
-				return frappe.get_meta(doctype)
-			except Exception:
-				return None
+		eval_locals.update(
+			{
+				"item": context.get("item"),
+				"loop": context.get("loop"),
+				"caller": frappe._dict(caller_meta or {}),
+				"rule": frappe._dict(rule_meta or {}),
+				"doctype": doctype_name,
+			}
+		)
 
-		def _is_submittable(doctype):
-			meta = _get_meta(doctype)
-			return bool(getattr(meta, "is_submittable", 0)) if meta else False
-
-		def _has_field(doctype, fieldname):
-			meta = _get_meta(doctype)
-			return bool(meta and fieldname and meta.has_field(fieldname))
-
-		def _length_of(value):
-			if value is None:
-				return 0
-			if isinstance(value, str):
-				return len(value.strip())
-			if isinstance(value, list | tuple | dict | set):
-				return len(value)
-			try:
-				return len(value)
-			except Exception:
-				return 0
-
-		def _is_empty_value(value):
-			if value is None:
-				return True
-			if isinstance(value, str):
-				return value.strip() == ""
-			if isinstance(value, list | tuple | dict | set):
-				return len(value) == 0
-			return False
-
-		return {
-			"doc": doc,
-			"old_doc": old_doc,
-			"vars": context.get("vars", {}),
-			"item": context.get("item"),
-			"loop": context.get("loop"),
-			"frappe": context.get("frappe", _safe_frappe),
-			"caller": frappe._dict(caller_meta or {}),
-			"rule": frappe._dict(rule_meta or {}),
-			"doctype": doctype_name,
-			"is_submittable": _is_submittable,
-			"has_field": _has_field,
-			"get_meta": _get_meta,
-			"resolve": FieldResolver.resolve,
-			"check_link_match": check_link_match,
-			"length_of": _length_of,
-			"is_empty_value": _is_empty_value,
-			"any": any,
-			"all": all,
-			"True": True,
-			"False": False,
-			"None": None,
-		}
+		return eval_locals
 
 	def _evaluate_python_condition(self, expression, context):
 		"""Evaluate Python expression safely and coerce to bool."""
@@ -861,9 +667,17 @@ class RuleEngine:
 			validation_result = result["result"]
 
 		# 1. Output Mapping (Result -> Context)
-		action_config_raw = frappe.parse_json(getattr(action, "config", "{}") or "{}")
-		action_config = action_config_raw if isinstance(action_config_raw, dict) else {}
-		output_mapping = action_config.get("output_mapping")
+		from flexirule.ruleflow.core.action_plan_cache import get_action_plan
+
+		plan = get_action_plan(self.rule, action)
+		output_mapping = plan.get("output_mapping") if plan else None
+
+		if not output_mapping and not plan:
+			# Fallback if plan is missing
+			action_config_raw = frappe.parse_json(getattr(action, "config", "{}") or "{}")
+			action_config = action_config_raw if isinstance(action_config_raw, dict) else {}
+			output_mapping = action_config.get("output_mapping")
+
 		if output_mapping:
 			if getattr(action, "is_async", 0):
 				raise MethodExecutionError(_("Async actions cannot map outputs"))
